+++ desktop/src/Umbrella.Wallet.Core/Chains/ChainId.cs
@@ -16,4 +16,5 @@ public enum ChainId
     Xmr,
     Bch,
     Zec,
+    Xrp,
 }